@@ -44,15 +44,16 @@
     "firebase-admin": "^13.4.0",
     "firebase-functions": "^6.4.0",
     "jest": "^30.1.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
     "react-native": "0.79.5",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-web": "~0.20.0",
-    "react-native-webview": "13.13.5"
+    "react-native-webview": "13.13.5",
+    "react-test-renderer": "^19.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
@@ -65,6 +66,7 @@
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
     "jest-expo": "^53.0.0",
+    "node-fetch": "^3.3.2",
     "typescript": "~5.8.3"
   },
   "private": true,
